rme: ath25: name globe led as "etactica"
Earlier, we hardcoded the led names to match the dragino ms12 platform.
Now, insert the uci system labels for the globe led as "etactica" as it
is on our cases.

Ensure that the "default" value is unrecognised by the init.d scripts,
so that the led is _registered_ by name, but no action is taken.
(scripts in the boot process are already taking decisions on the led
state by the time the 96_leds script runs)